fix: 축의금 납부 목록 조회 기능의 memberWedding 폴더 삭제 후 wedding 폴더에 옮김
