+++ src/main/java/com/kestrel/weddingbookkeeper/api/wedding/controller/WeddingController.java
@@ -1,26 +1,24 @@
 package com.kestrel.weddingbookkeeper.api.wedding.controller;
 
+import com.kestrel.weddingbookkeeper.api.wedding.vo.MemberWedding;
 import com.kestrel.weddingbookkeeper.api.wedding.dto.request.PartnerCodeRequest;
 import com.kestrel.weddingbookkeeper.api.wedding.dto.response.WeddingInfoResponse;
 import com.kestrel.weddingbookkeeper.api.wedding.facade.WeddingFacade;
-import com.kestrel.weddingbookkeeper.api.wedding.dto.WeddingInfoRequestDto;
 import com.kestrel.weddingbookkeeper.api.wedding.dto.request.WeddingInfoRequest;
 import com.kestrel.weddingbookkeeper.api.wedding.service.WeddingService;
-import com.kestrel.weddingbookkeeper.api.wedding.vo.Wedding;
-import org.springframework.ui.Model;
+import java.util.List;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/v1/weddings")
 public class WeddingController {
 
-    private static final Integer MEMBER_ID = 7;
+    private static final Integer MEMBER_ID = 1;
 
     private final WeddingFacade weddingFacade;
     private final WeddingService weddingService;
@@ -45,5 +43,11 @@ public class WeddingController {
         System.out.println("partnerCodeRequest = " + partnerCodeRequest);
         boolean isConnect = weddingService.connectPartner(partnerCodeRequest, MEMBER_ID);
         System.out.println("isConnect = " + isConnect);
+    }
+
+    @GetMapping
+    public List<MemberWedding> selectDonationList() {
+        List<MemberWedding> memberWeddingList = weddingService.selectDonationList(MEMBER_ID);
+        return memberWeddingList;
     }
 }
