+++ src/main/resources/mapper/MemberWeddingMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.kestrel.weddingbookkeeper.api.memberWedding.dao.MemberWeddingDao">
+<mapper namespace="com.kestrel.weddingbookkeeper.api.wedding.dao.MemberWeddingDao">
 
   <resultMap id="memberMap" type="com.kestrel.weddingbookkeeper.api.member.vo.Member">
     <result column="member_id" property="memberId"/>
